[NEW] add page 404 and function recherche

+++ src/Controller/CommentaireController.php
@@ -45,6 +45,11 @@ class CommentaireController extends AbstractController
         if (!$commentaire) {
             throw $this->createNotFoundException('Commentaire non trouvé');
         }
+        // Restriction : seul l'auteur ou un admin peut supprimer
+        $user = $this->getUser();
+        if (!$user || ($user !== $commentaire->getUser() && !in_array('ROLE_ADMIN', $user->getRoles()))) {
+            throw $this->createAccessDeniedException('Vous ne pouvez pas supprimer ce commentaire.');
+        }
         $videoId = $commentaire->getResumeVideo()->getId();
         $em->remove($commentaire);
         $em->flush();
